Basic_OOP_with_Java_v_0.2
Exercise introduce concept of a Interface. Interface is a fully abstract class containing only a abstract methods, it means that it create "contract" for it implementator to code them. Our interface is responsible for saving and recovering main School object in serialized format. For ability to save serialized object we must implement interface Serializable to our class School and all participating classes. Interface serializable is a only marking interface that does not make us responsible for implementation of any further method. For better understanding of concept of the interface we also created BlueJ version of this exercise.