Basic_OOP_with_Java_v_0.5
cv_4_ introduce example of abstract class and final method marking. Abstract class cannot to be instantionated and act as a template for implementing subcasses. Final methot can not to be overriden in child class. Reason for do this can be seen in protecting our code from unauthorized altering. Most of java API functions are marked as final.